+++ sdk/eventhub/eventhubs-checkpointstore-blob/changelog.md
@@ -1,6 +1,14 @@
+### 2019-11-05 - 1.0.0-preview.5
+
+- Updated to use the latest version of the `@azure/event-hubs` package.
+
+Breaking changes:
+- `BlobPartitionManager` has been renamed to `BlobCheckpointStore` to reflect naming changes 
+   made in the `@azure/event-hubs` package.
+
 ### 2019-11-05 - 1.0.0-preview.4
 
-- Updated to use the latest version of the `@azure/event-hubs` package and `@azure/storabe-blob` packages.
+- Updated to use the latest version of the `@azure/event-hubs` package and `@azure/storage-blob` packages.
 
 ### 2019-10-08 1.0.0-preview.3